requirements.txt
tensorflow; sys_platform != 'darwin'
tensorflow-macos; sys_platform == 'darwin'
pandas
numpy
matplotlib
seaborn
scikit-learn
jupyter
streamlit
pillow
